refactor: just use serde instead
